app/requirements.txt
# requirements.txt - Contains all required packages for the HealthEconomics360 platform

Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.0.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
Flask-Bcrypt==0.7.1
psycopg2-binary==2.9.1
pandas==1.3.0
numpy==1.21.0
scikit-learn==0.24.2
plotly==5.1.0
dash==2.0.0
gunicorn==20.1.0
python-dotenv==0.19.0
email-validator==1.1.3
pytest==6.2.5
Flask-RESTful==0.3.9
Jinja2==3.0.1
scipy==1.7.0
Werkzeug==2.0.1